Making the tests 'work' in separate folder
